+++ src/main/java/com/pkielbasa/pocketplan/api/dto/transaction/CreateTransactionRequest.java
@@ -1,10 +1,9 @@
-package com.pkielbasa.pocketplan.application.dto.transaction;
+package com.pkielbasa.pocketplan.api.dto.transaction;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-public record TransactionResponse(
-        Long id,
+public record CreateTransactionRequest(
         Long budgetId,
         String name,
         String description,
