+++ src/main/java/com/pkielbasa/pocketplan/api/dto/user/CreateUserRequest.java
@@ -1,4 +1,4 @@
-package com.pkielbasa.pocketplan.application.dto.user;
+package com.pkielbasa.pocketplan.api.dto.user;
 
 public record CreateUserRequest(
         String username,
