+++ src/main/java/com/pkielbasa/pocketplan/api/dto/user/UserResponse.java
@@ -1,3 +1,3 @@
-package com.pkielbasa.pocketplan.application.dto.user;
+package com.pkielbasa.pocketplan.api.dto.user;
 
 public record UserResponse(Long id, String username, String email) {}
