+++ src/main/java/com/pkielbasa/pocketplan/api/rest/TransactionController.java
@@ -1,8 +1,8 @@
 package com.pkielbasa.pocketplan.api.rest;
 
 import com.pkielbasa.pocketplan.api.mapper.TransactionMapper;
-import com.pkielbasa.pocketplan.application.dto.transaction.CreateTransactionRequest;
-import com.pkielbasa.pocketplan.application.dto.transaction.TransactionResponse;
+import com.pkielbasa.pocketplan.api.dto.transaction.CreateTransactionRequest;
+import com.pkielbasa.pocketplan.api.dto.transaction.TransactionResponse;
 import com.pkielbasa.pocketplan.application.exception.ResourceNotFoundException;
 import com.pkielbasa.pocketplan.application.service.TransactionService;
 import com.pkielbasa.pocketplan.domain.model.Transaction;
